next-client-over

+++ Client/utils/localStorageMessage.js
@@ -2,7 +2,7 @@ class LocalStorageMessage{
     /** 获取localstorage数据 **/
     getLocalStorage = key => {
         try {
-            return JSON.parse(localStorage.getItem(key));
+            return localStorage && JSON.parse(localStorage.getItem(key));
         }catch (error){
             console.log('error',error)
         }
@@ -11,7 +11,7 @@ class LocalStorageMessage{
     /** 设置localstorage数据 **/
     setLocalStorage = (key, params) => {
         try {
-            localStorage.setItem(key, JSON.stringify(params));
+            localStorage && localStorage.setItem(key, JSON.stringify(params));
         }catch (error){
             console.log('error',error)
         }
@@ -20,7 +20,7 @@ class LocalStorageMessage{
     /** 删除localstorage **/
     removeLocalStorage = key => {
         try {
-            localStorage.removeItem(key);
+            localStorage && localStorage.removeItem(key);
         }catch (error){
             console.log('error',error)
         }
@@ -29,7 +29,7 @@ class LocalStorageMessage{
     /** 清除localstorage **/
     clearLocalStorage = () => {
         try {
-            localStorage.clear();
+            localStorage && localStorage.clear();
         }catch (error){
             console.log('error',error)
         }
